requirements.txt
numba
numpy
pyfunctional